pyproject.toml
[tool.poetry]
name = "parkings"
version = "0.1.0"
description = ""
authors = ["uventus <uventus_work@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
uvicorn = "^0.32.0"
yookassa = "^3.3.0"
wrapt = "^1.16.0"
fastapi-pagination = "^0.12.31"
sqlalchemy = "^2.0.36"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.0"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
python-dateutil = "^2.9.0.post0"
python-multipart = "^0.0.13"
databases = "^0.9.0"
greenlet = "^3.1.1"
pyjwt = "^2.9.0"
fastapi-amis-admin = "^0.7.2"
fastapi = "0.103.2"
bcrypt = "4.0.0"
fastapi-user-auth = "^0.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
